requirements.txt
altgraph==0.17.4
bidict==0.23.1
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-engineio==4.12.3
python-socketio==5.14.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
shiboken6==6.10.0
simple-websocket==1.1.0
snakeviz==2.2.2
tornado==6.5.2
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
zipp==3.23.0
